Book show 50147

+++ db/migrate/20160104125344_add_rating_to_books.rb
@@ -1,5 +1,5 @@
 class AddRatingToBooks < ActiveRecord::Migration
   def change
-    add_column :books, :rating, :integer, default: 0
+    add_column :books, :rating, :float, default: 0
   end
 end
